change address type
increase from `unsigned int` (4 bytes) to `size_t` (8 bytes)

+++ tls.c
@@ -33,8 +33,8 @@ typedef struct thread_local_storage {
 } TLS;
 
 struct page {
-  unsigned int address; /* start address of page */
-  int ref_count;        /* counter for shared pages */
+  size_t address; /* start address of page */
+  int ref_count;  /* counter for shared pages */
 };
 
 ENTRY e, *ep;
@@ -124,14 +124,14 @@ static void tls_init() {
 }
 
 static void tls_protect(struct page *p) {
-  if (mprotect((void *)(size_t)p->address, page_size, PROT_NONE)) {
+  if (mprotect((void *)p->address, page_size, PROT_NONE)) {
     fprintf(stderr, "tls_protect: could not protect page\n");
     exit(EXIT_FAILURE);
   }
 }
 
 static void tls_unprotect(struct page *p) {
-  if (mprotect((void *)(size_t)p->address, page_size, PROT_READ | PROT_WRITE)) {
+  if (mprotect((void *)p->address, page_size, PROT_READ | PROT_WRITE)) {
     fprintf(stderr, "tls_unprotect: could not unprotect page\n");
     exit(EXIT_FAILURE);
   }
@@ -166,7 +166,7 @@ static struct page *create_copy(struct page *p) {
   copy->address = (size_t)mmap(0, page_size, PROT_WRITE,
                                MAP_ANONYMOUS | MAP_PRIVATE, -1, 0);
   copy->ref_count = 1;
-  memcpy((void *)(size_t)copy->address, (void *)(size_t)p->address, page_size);
+  memcpy((void *)copy->address, (void *)p->address, page_size);
   p->ref_count--;
   tls_protect(p);
   return copy;
@@ -225,7 +225,7 @@ int tls_destroy() {
   for (int i = 0; i < lsa->num_pages; i++) {
     struct page *p = lsa->pages[i];
     if (--p->ref_count == 0) {
-      munmap((void *)(size_t)p->address, page_size);
+      munmap((void *)p->address, page_size);
       free(p);
     }
   }
@@ -249,7 +249,7 @@ int tls_read(unsigned int offset, unsigned int length, char *buffer) {
   unsigned int cnt, idx;
   for (cnt = 0, idx = offset; idx < (offset + length); ++cnt, ++idx) {
     struct page *p = lsa->pages[idx / page_size];
-    buffer[cnt] = *((char *)(size_t)(p->address + idx % page_size));
+    buffer[cnt] = *((char *)(p->address + idx % page_size));
   }
 
   for (int i = 0; i < lsa->num_pages; i++)
@@ -274,7 +274,7 @@ int tls_write(unsigned int offset, unsigned int length, const char *buffer) {
     if (p->ref_count > 1) {
       p = lsa->pages[idx / page_size] = create_copy(p);
     }
-    *((char *)(size_t)(p->address + idx % page_size)) = buffer[cnt];
+    *((char *)(p->address + idx % page_size)) = buffer[cnt];
   }
 
   for (int i = 0; i < lsa->num_pages; i++)
